create_close supports noverify to bypass expensive checks
A trusted client may send "noverify=1" and a proper size for the
tracker to skip size verification.  "noverify" also overrides
the "checksumverify" option (if set).

Size verification can be expensive on high-latency routes
between the tracker and storage node used, so this allows
queryworkers to work more quickly on file injection.

Since the MogileFS already places a high amount of trust in
clients, the use of this option is unlikely to cause problems
for non-buggy clients.  The default (size verification enabled)
remains unchanged.