add conn_pool_size configuration option
This defines the size of the HTTP connection pool.  This affects
all workers at the moment, but is likely most interesting to the
Monitor as it affects the number of devices the monitor may
concurrently update.

This defaults to 20 (the long-existing, hard-coded value).

In the future, there may be a(n easy) way to specify this on a
a per-worker basis, but for now it affects all workers.